Move out of args::lockfile
