+++ cli/args/mod.rs
@@ -4,7 +4,6 @@ pub mod deno_json;
 mod flags;
 mod flags_net;
 mod import_map;
-mod lockfile;
 pub mod package_json;
 
 pub use self::import_map::resolve_import_map;
@@ -34,9 +33,6 @@ pub use deno_config::TsConfigType;
 pub use deno_config::TsTypeLib;
 pub use deno_config::WorkspaceConfig;
 pub use flags::*;
-pub use lockfile::read_lockfile_at_path;
-pub use lockfile::write_lockfile_if_has_changes;
-pub use lockfile::Lockfile;
 pub use package_json::PackageJsonDepsProvider;
 
 use deno_ast::ModuleSpecifier;
@@ -74,6 +70,7 @@ use thiserror::Error;
 use crate::args::import_map::enhance_import_map_value_with_workspace_members;
 use crate::cache;
 use crate::file_fetcher::FileFetcher;
+use crate::lockfile::Lockfile;
 use crate::util::fs::canonicalize_path_maybe_not_exists;
 use crate::version;
 
@@ -958,7 +955,7 @@ impl CliOptions {
       }),
     )?;
 
-    let maybe_lock_file = lockfile::discover(
+    let maybe_lock_file = crate::lockfile::discover(
       &flags,
       maybe_config_file.as_ref(),
       maybe_package_json.as_deref(),
